Cards are deleted in contentful when an activity is deleted in the database

+++ backend/activities/routes.py
@@ -2,7 +2,7 @@ from flask import (Blueprint, request)
 from flask_restful import Resource
 from backend import api, db
 from backend.activities.schemas import activity_schema, activities_schema
-from backend.activities.utils import create_activity, edit_activity
+from backend.activities.utils import create_activity, delete_cards, edit_activity
 from backend.models import Activity
 
 # Blueprint for activities
@@ -30,7 +30,12 @@ class ActivityCRUD(Resource):
     # Function to edit an activity
     def put(self):
         contentful_data = request.get_json()
-        edit_activity(contentful_data)
+        activity = Activity.query.filter_by(contentful_id=contentful_data["entityId"]).first()
+
+        if not activity:
+            return {"message": "Activity does not exist"}, 404
+
+        edit_activity(activity, contentful_data)
 
         db.session.commit()
 
@@ -43,6 +48,11 @@ class ActivityDelete(Resource):
     def post(self):
         contentful_data = request.get_json()
         activity = Activity.query.filter_by(contentful_id=contentful_data["entityId"]).first()
+
+        if not activity:
+            return {"message": "Activity does not exist"}, 404
+
+        delete_cards(activity.cards)
 
         db.session.delete(activity)
         db.session.commit()
